Fix two portability defects the first CI run found
Neither was reachable from this bench: the library had only ever been built with
gcc on Linux, and CI had no remote to run on until now. Both were fatal under
-Werror / /WX.

⚠ `NAN` is a FLOAT constant and `hm_report`'s medians are all double, so every
use implicitly widened. clang rejects that under -Wdouble-promotion — which this
project has enabled for both compilers — while gcc does not warn. Five sites in
`record/hm_reconcile.c`, now cast explicitly. This broke linux-clang, linux-san
and macos identically.

⚠ MSVC C4701 on `history_capture_fits`: it cannot see that `have` guarantees the
snapshot is taken before `snap.flags` is read, and /WX makes the warning fatal.
Zero-initialised, with the reasoning at the declaration. The zero value is the
safe one if that flow ever changes — flags without HM_CLOCK_HAS_FIT makes the
branch return rather than proceed on an empty fit.

⚠ Neither changes behaviour and neither touches §6.1.1's re-anchor, which
`history_capture_fits` is one of the three homes of. gcc build, sanitizer build,
24 tests each and both fixture replays are unchanged; the clang and MSVC halves
can only be confirmed by CI, which is the point of having it.

+++ record/hm_reconcile.c
@@ -791,11 +791,15 @@ static int cmp_i16(const void *a, const void *b)
     return (x < y) ? -1 : ((x > y) ? 1 : 0);
 }
 
-/* Median of a copy; `values` is not disturbed. */
+/* Median of a copy; `values` is not disturbed.
+ *
+ * ⚠ NAN is a FLOAT constant and every target here is a double, so each use is
+ * cast explicitly.  clang's -Wdouble-promotion rejects the implicit widening
+ * (gcc does not, which is why this only surfaced in CI). */
 static double median_of(const int16_t *values, size_t n, int16_t *scratch)
 {
     if (n == 0u) {
-        return NAN;
+        return (double)NAN;
     }
     memcpy(scratch, values, n * sizeof(int16_t));
     qsort(scratch, n, sizeof(int16_t), cmp_i16);
@@ -810,10 +814,10 @@ static void finish_skew(hm_reconciler *r)
     size_t n = (size_t)r->rep.skew_stored;
     int16_t *scratch;
 
-    r->rep.skew_median_ticks = NAN;
-    r->rep.skew_median_first_half = NAN;
-    r->rep.skew_median_second_half = NAN;
-    r->rep.skew_median_us = NAN;
+    r->rep.skew_median_ticks = (double)NAN;
+    r->rep.skew_median_first_half = (double)NAN;
+    r->rep.skew_median_second_half = (double)NAN;
+    r->rep.skew_median_us = (double)NAN;
     if (n == 0u) {
         return;
     }
